# api-gateway/src/docs/openapi3_0.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Supermarket API",
    "version": "1.0.0",
    "description": "A Node.js Express API for managing supermarket-related services including authentication, users, products, and purchases. The system is designed with scalability and maintainability in mind, separating concerns into independent services that follow the  Model-View-Controller(MVC) design pattern and communicate through an API Gateway."
  },
  "servers": [
    {
      "url": "http://localhost:3000/supermarket/",
      "description": "Base URL for the supermarket API"
    }
  ],
...
  "paths": {
...
    "/auth/verify-token": {
      "post": {
        "tags": [
          "authenticate"
        ],
        "summary": "Verify Tokens.",
        "operationId": "verifyToken",
        "description": "Verifies the access tokens provided in the request and refreshing them if the access token is expired.",
        "security": [
          {
            "AccessToken": []
          },
          {
            "RefreshToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Tokens verified successfully. Returns user details and optionally new tokens if refreshed.",
            "headers": {
              "Authorization": {
                "schema": {
                  "type": "string",
                  "description": "Access token returned in the Authorization header."
                }
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string",
                  "description": "Refresh token set in the `refresh_token` cookie."
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Tokens verified successfully."
                    },
                    "user": {
                      "$ref": "#/components/schemas/UserResponse"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized due to missing or invalid secret key or token issues.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/UnauthorizedResponse"
                    },
                    {
                      "$ref": "#/components/schemas/InvalidSecretKey"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh-token": {
      "post": {
        "tags": [
          "authenticate"
        ],
        "summary": "Refresh Tokens.",
        "operationId": "refreshToken",
        "description": "Refreshes the access and refresh tokens using a valid refresh token.",
        "security": [
          {
            "RefreshToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Tokens refreshed successfully. Returns user details and new tokens.",
            "headers": {
              "Authorization": {
                "schema": {
                  "type": "string",
                  "description": "Access token returned in the Authorization header."
                }
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string",
                  "description": "Refresh token set in the `refresh_token` cookie."
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Tokens refreshed successfully."
                    },
                    "user": {
                      "$ref": "#/components/schemas/UserResponse"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized due to missing or invalid secret key or token issues.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/UnauthorizedResponse"
                    },
                    {
                      "$ref": "#/components/schemas/InvalidSecretKey"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier of the user.",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "Michael Kenneth Williams"
          },
          "username": {
            "type": "string",
            "description": "Username chosen by the user for login or identification purposes.",
            "example": "mkwilliams"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the user, used for communication.",
            "example": "michael.williams@x.super.com\""
          },
          "role": {
            "type": "string",
            "description": "Role assigned to the user, defining their permissions and      access level.",
            "enum": [
              "admin",
              "manager",
              "supervisor",
              "user"
            ],
            "example": "USER"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp indicating when the user account was created.",
            "example": "2025-03-15T10:30:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp indicating when the user account was last updated.",
            "example": "2025-03-25T11:56:00Z"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": [
          "message"
        ]
      },
      "InvalidSecretKey": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Message describing the error",
            "example": "Unauthorized. Missing or invalid secret-key."
          },
          "errorCode": {
            "type": "string",
            "description": "Specific error code",
            "example": "InvalidSecretKey"
          }
        },
        "required": [
          "message",
          "errorCode"
        ]
      },
...
      "UnauthorizedResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "You are not authorized to perform this action."
          },
          "code": {
            "type": "string",
            "example": "InvalidToken"
          }
        }
      },
...
    },
    "securitySchemes": {
      "AccessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "in": "header",
        "name": "authorization",
        "description": "Access token passed in the Authorization header"
      },
      "RefreshToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "in": "cookie",
        "name": "refresh_token",
        "description": "Refresh token passed as refresh_token=bearer<JWT> in the Cookie header"
      }
    }
  }
}